# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Provide user with CRUD operations with DB of events/tasks/notes",
    "version": "1.0.0",
    "title": "Notes API"
  },
  "schemes": ["http"],
  "host": "localhost:3000",
  "basePath": "/docs",
  "paths" : {
    "/api/v1/entity" : {
      "get" : {
        "summary" : "Get all the tasks",
        "description": "Get all the tasks",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/todosResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/api/v1/entity/{id}" : {
      "get" : {
        "summary" : "Get all the tasks",
        "description": "Get all the tasks",
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/todosResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/api/v1/entity/{id}": {
      "post": {
        "summary": "Add an instance of specific object with generated id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The unique id of the entity",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "type",
                  "title",
                  "content",
                  "date",
                  "time"
                ],
                "properties": {
                  "id": {
                    "type": "integer",
                    "default": 2
                  },
                  "type": {
                    "type": "string",
                    "default": "task"
                  },
                  "title": {
                    "type": "string",
                    "default": "Some Title"
                  },
                  "content": {
                    "type": "string",
                    "default": "Some Description"
                  },
                  "date": {
                    "type": "string",
                    "default": "May 08, 2023"
                  },
                  "time": {
                    "type": "string",
                    "default": "12:49 AM"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "default": 0
                    },
                    "type": {
                      "type": "string",
                      "default": "task"
                    },
                    "title": {
                      "type": "string",
                      "default": "Do it NOW!"
                    },
                    "content": {
                      "type": "string",
                      "default": "just a kind reminder :)"
                    },
                    "date": {
                      "type": "string",
                      "default": "May 08, 2023"
                    },
                    "time": {
                      "type": "string",
                      "default": "12:49 AM"
                    },
                    "dueDate": {
                      "type": "string",
                      "default": "2023-05-09"
                    },
                    "isCompleted": {
                      "type": "boolean",
                      "default": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Title and description are required"
          }
        }
      }
    },
    "/todos/{id}" : {
      "put" : {
        "summary" : "Update the tasks",
        "description": "Update the tasks",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id that needs to be deleted",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "task object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "task" : {
                  "type": "object",
                  "$ref": "#/definitions/Task"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/todosResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    },
    "/todo/{id}" : {
      "delete" : {
        "summary" : "Delete the task",
        "description": "Delete the task",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "task id that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/todosResponse"
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "schema": {
              "$ref": "#/definitions/InvalidResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "todosResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "task": {
          "type": "string"
        },
        "assignee": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "Task": {
      "type": "object",
      "properties": {
        "task": {
          "type": "string"
        },
        "assignee": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "InvalidResponse": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }

    }
  }
}